use vm again

+++ test/runWebpack.js
@@ -13,14 +13,18 @@ const config = {
   },
 }
 
-module.exports = plugin => {
-  fse.removeSync(plugin.opts.fullFilename)
-  fse.removeSync(`${plugin.opts.fullFilename}.map`)
+module.exports = plugins => {
+  if (!Array.isArray(plugins)) plugins = [plugins]
+  plugins.forEach(plugin => {
+    if (!plugin.opts || plugin.opts.fullFilename) return
+    fse.removeSync(plugin.opts.fullFilename)
+    fse.removeSync(`${plugin.opts.fullFilename}.map`)
+  })
 
   const compiler = webpack(
     Object.assign(
       {
-        plugins: [plugin],
+        plugins,
       },
       config
     )
